move all config files in batch to a separate folder
