+++ dfastmi/config/AConfigurationChecker.py
