+++ dfastmi/gui/dialog_view_model.py
@@ -33,8 +33,8 @@ from typing import Dict
 from PyQt5.QtCore import QObject, pyqtSignal
 
 import dfastmi
-from dfastmi.batch.ConfigurationInitializer import ConfigurationInitializer
 from dfastmi.batch.DFastUtils import get_progloc
+from dfastmi.config.ConfigurationInitializer import ConfigurationInitializer
 from dfastmi.gui.dialog_model import DialogModel
 from dfastmi.io.ApplicationSettingsHelper import ApplicationSettingsHelper
 from dfastmi.io.AReach import AReach
