+++ tests/batch/test_AnalyserDflowfm.py
@@ -6,10 +6,10 @@ import shapely
 from mock import Mock, patch
 from shapely.geometry.linestring import LineString
 
-from dfastmi.batch.AConfigurationInitializerBase import AConfigurationInitializerBase
 from dfastmi.batch.AnalyserDflowfm import AnalyserDflowfm
 from dfastmi.batch.SedimentationData import SedimentationData
 from dfastmi.batch.XykmData import XykmData
+from dfastmi.config.AConfigurationInitializerBase import AConfigurationInitializerBase
 from dfastmi.io.map_file import MapFile
 from tests.batch.Helper_AnalyserAndReporterDflowfm import (  # needed for fixture
     TestCase_display_needs_tide_old_zmin_zmax,
