+++ tests/config/test_DataAccess_core.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 
 import pytest
 
-from dfastmi.batch.ConfigurationCheckerLegacy import DFLOWFM_MAP, WAQUA_EXPORT
+from dfastmi.config.ConfigurationCheckerLegacy import DFLOWFM_MAP, WAQUA_EXPORT
 from dfastmi.io.ConfigFileOperations import ConfigFileOperations, check_configuration
 from dfastmi.io.RiversObject import RiversObject
